VV 'fake' setters, fixes the visualizer, fixes the singularity trying to eat itself instead of nearby things.

+++ Content.Client/Singularity/Visualizers/SingularityVisualizer.cs
@@ -32,7 +32,7 @@ namespace Content.Client.Singularity.Visualizers
                 return;
             }
 
-            if (!component.TryGetData(SingularityVisuals.Level, out int level))
+            if (!component.TryGetData(SingularityVisuals.Level, out byte level))
             {
                 return;
             }
